[BACKLOG-6051] Move validate from Value to Item
